Add mima binary compatibility plugin
Now `sbt/mimaReportBinaryIssues` should fail if there is a change that
isn't binary compatible in the core, http, or zeromq module.

In the search client release Jenkins job I have added a Boolean build
parameter for whether or not to check binary compatibility. It defaults
to checked and I uncheck it if I have just bumped the version number for
a release that I'm okay with not being binary compatible. I leave it up
to you how you want to handle changes to the Jenkins job for funnel.