Make Metric a sealed trait
Currently it does a case-match that isn't guaranteed to be exhaustive.

From a binary-compatibility standpoint, it should probably be a sealed
abstract class. That way, methods could be added to it without breaking
binary compatibility. However, changing it from a trait to an abstract
class would itself be a binary-incompatible change. It may be something
to consider for the next incompatible version of funnel though (I
imagine there are other traits that could do the same).